Change fork/accept/event model
Instead of events, A is a command that is run to accept a socket. You then
fork to process the socket and close it in the main thread.

A bit unfortunate though - it would've been nice to hide the accepting in
the runtime so it can be parallelized without explicitly writing that in
the xed program.